requirement.txt
streamlit>=1.33.0
fastapi
uvicorn[standard]
httpx
pydantic-settings
requests
python-dotenv